requirements.txt
numpy
keras
matplotlib
opencv-python
tensorflow
scikit-learn
Pillow
seaborn